requirements.txt
slack-bolt>=1.18.0
slack-sdk>=3.26.0
fastapi>=0.110.0
uvicorn>=0.29.0
httpx>=0.27.0
python-dotenv>=1.0.0
